requirements.txt
python-telegram-bot==21.4
fastapi==0.115.0
uvicorn==0.30.6
anthropic==0.39.0
python-dotenv==1.0.1
aiosqlite==0.20.0
